refactor(src): create utils folder
